+++ test/test_physics.py
@@ -92,7 +92,7 @@ class TestPhysics(EnvironmentSetup):
             env.StepSimulation(0.01)
         with env:
             T3 = body.GetTransform()
-            print (T[2,3]-Tinit[2,3]),(T3[2,3]-T2[2,3])
+            self.log.info('differences: %f, %f', (T[2,3]-Tinit[2,3]),(T3[2,3]-T2[2,3]))
             assert( abs((T[2,3]-Tinit[2,3]) - (T3[2,3]-T2[2,3])) < 0.001)
         assert(abs(1e-6*env.GetSimulationTime()-simtime1-(simtime0-starttime)) < g_epsilon)