+++ test/test_planning.py
@@ -61,7 +61,7 @@ class TestMoving(EnvironmentSetup):
         env.StartSimulation(0.01,False)
         robot.WaitForController(0)
         with env:
-            print 'test_ikplanning:',transdist(Tee,ikmodel.manip.GetEndEffectorTransform())
+            self.log.info('Tee dist=%f',transdist(Tee,ikmodel.manip.GetEndEffectorTransform()))
             assert(transdist(Tee,ikmodel.manip.GetEndEffectorTransform()) <= g_epsilon)
             
     def test_constraintpr2(self):
@@ -85,18 +85,18 @@ class TestMoving(EnvironmentSetup):
         self.LoadEnv('data/lab1.env.xml')
         robot=env.GetRobots()[0]
 
-        print 'test_constraintwam: ikmodel'
+        self.log.debug('generating ikmodel')
         ikmodel = databases.inversekinematics.InverseKinematicsModel(robot=robot,iktype=IkParameterization.Type.Transform6D)
         if not ikmodel.load():
             ikmodel.autogenerate()
-        print 'test_constraintwam: gmodel'
+        self.log.debug('generating gmodel')
         gmodel = databases.grasping.GraspingModel(robot=robot,target=env.GetKinBody('mug1'))
         if not gmodel.load():
             # don't do multithreaded yet since ode on some ubuntu distors does not support it
             gmodel.numthreads = 2 # at least two threads
             gmodel.generate(approachrays=gmodel.computeBoxApproachRays(delta=0.04))
             gmodel.save()
-        print 'test_constraintwam: planning'
+        self.log.debug('planning')
         with env:
             basemanip = interfaces.BaseManipulation(robot)
             robot.SetActiveDOFs(ikmodel.manip.GetArmIndices())
@@ -119,7 +119,7 @@ class TestMoving(EnvironmentSetup):
             assert(solgoal is not None)
             traj = basemanip.MoveToHandPosition(matrices=[Tgoal],maxiter=3000,maxtries=1,seedik=40,constraintfreedoms=constraintfreedoms,constraintmatrix=constraintmatrix,constrainttaskmatrix=constrainttaskmatrix,constrainterrorthresh=constrainterrorthresh,steplength=0.002,outputtrajobj=True,execute=False,jitter=0.05)
             soltraj = traj.Sample(0,robot.GetActiveConfigurationSpecification())
-            print 'test_constraintwam: make sure it starts at the initial configuration'
+            self.log.debug('make sure it starts at the initial configuration')
             assert(transdist(soltraj,sol) <= g_epsilon)
             self.RunTrajectory(robot,traj)
                 
@@ -160,7 +160,7 @@ class TestMoving(EnvironmentSetup):
             sol = robot.GetActiveDOFValues()
             traj = basemanip.MoveManipulator(orgvalues,outputtrajobj=True,execute=False,jitter=0.05)
             soltraj = traj.Sample(0,robot.GetActiveConfigurationSpecification())
-            print 'test_wamgraspfromcollision: make sure it starts at the initial configuration'
+            self.log.debug('make sure it starts at the initial configuration')
             assert(transdist(soltraj,sol) <= g_epsilon)
             self.RunTrajectory(robot,traj)
                 
@@ -306,7 +306,7 @@ class TestMoving(EnvironmentSetup):
             robot.SetDOFValues([0.187],[robot.GetJoint('l_shoulder_lift_joint').GetDOFIndex()])
             assert(env.CheckCollision(robot))
             
-            print 'test_planwithcollision: environment collision'
+            self.log.debug('environment collision')
             ikmodel = databases.inversekinematics.InverseKinematicsModel(robot,iktype=IkParameterization.Type.Transform6D)
             if not ikmodel.load():
                 ikmodel.autogenerate()
@@ -319,7 +319,7 @@ class TestMoving(EnvironmentSetup):
             ret = basemanip.MoveToHandPosition([Tnew],execute=False)
             assert(ret is not None)
 
-            print 'test_planwithcollision: self collision'
+            self.log.debug('self collision')
             robot.SetDOFValues(defaultvalues)
             robot.SetDOFValues([ 1.34046301,  0.94535038,  3.03934583, -1.30743665, 0 , 0 ,  0], robot.GetManipulator('leftarm').GetArmIndices())
             assert(robot.CheckSelfCollision())